requirements.txt
# Core data science (updated for Python 3.12)
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Advanced ML
xgboost>=1.7.0
optuna>=3.0.0

# MLOps tools
mlflow>=2.0.0
joblib>=1.2.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Testing & quality
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Additional needed packages
setuptools>=65.0.0
wheel>=0.38.0
requests>=2.31.0
python-multipart>=0.0.6